+++ src/analyze/+mag/@IMAPTestingAnalysis/loadIALiRTData.m
@@ -60,6 +60,12 @@ function loadIALiRTData(this, primaryMetaData, secondaryMetaData)
         secondaryData = vertcat(secondaryData, table2timetable(secondary, RowTimes = "t")); %#ok<AGROW>
     end
 
+    % Add continuity information, for simpler interpolation.
+    % Property order:
+    %     sequence, x, y, z, range, coarse, fine, compression, quality
+    [primaryData.Properties.VariableContinuity, secondaryData.Properties.VariableContinuity] = ...
+        deal(["step", "continuous", "continuous", "continuous", "step", "continuous", "continuous", "step", "step"]);
+
     %% Amend Timestamp
 
     startTime = bounds(primaryData.t);